+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- ``dtool report`` command for generating reports on collections of datasets
+
 
 Changed
 ^^^^^^^