+++ CHANGELOG.rst
@@ -26,11 +26,18 @@ Removed
 Fixed
 ^^^^^
 
-- Defect where inventory html template is not included in Python package on PyPi
-
 
 Security
 ^^^^^^^^
+
+
+[0.10.1] - 2018-05-18
+---------------------
+
+Fixed
+^^^^^
+
+- Defect where inventory html template is not included in Python package on PyPi
 
 
 [0.10.0] - 2018-05-18